Implement SSR for Lookup
