+++ testing/tests/DevExpress.serverSide/lookup.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/lookup.markup.tests.js");
